This is the patient finder content
I don't know where the "content" div is in the proper files. Please copy
& paste into the proper place. I have HTML and CSS in separate files
because I think this is preferred.

Commenting should explain everything in both files.
The "content" div tags and style can be deleted in both files. They are
just so you can see how they would look in the proper files.

This is also uploaded on the Facebook page.